# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SayurPintar API
  description: API for SayurPintar - helping street vegetable vendors optimize their daily operations
  version: 1.0.0
  contact:
    name: SayurPintar Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.sayurpintar.com/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AggregatedPrice:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        area:
          type: string
        median_price:
          type: number
          format: double
        min_price:
          type: number
          format: double
        max_price:
          type: number
          format: double
        sample_size:
          type: integer
        unit:
          type: string
        trend:
          type: string
          enum: [up, down, stable]
        change_pct:
          type: number
          format: double
        date:
          type: string
...
paths:
...
  /prices/stats/area:
    get:
      summary: Get price statistics for all products in an area
      tags: [Prices]
      parameters:
        - name: area
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
          description: Date in YYYY-MM-DD format (defaults to today)
      responses:
        "200":
          description: Area price statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AggregatedPrice"
  /prices/top-movers:
    get:
      summary: Get products with the biggest price changes
      tags: [Prices]
      parameters:
        - name: area
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        "200":
          description: Top movers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AggregatedPrice"

  # ── Product Endpoints ──
...